requirements.txt
pandas
numpy
scikit-learn
google-generativeai
streamlit
python-dotenv
tqdm
